Update ACS-Data-Analyst-GPT-Instructions
Specified that it was an ACS data expert, added data analysis handling and data sourcing. Also added defaulting to most recent data, if it exists, and prompting for more information as needed. 

Reason for these changes:
We are trying to correct the behavior of creating a query with (for example) using the 5year ACS data when the columns only existed in the 1year file. But also to hopefully provide a convenient url to get the table metadata to help validate code.